audit: product list section
