+++ apps/web/vibes/soul/examples.ts
@@ -1017,25 +1017,25 @@ export const examples = [
     component: lazy(() => import('./examples/sections/product-list/luxury')),
   },
   {
-    name: 'products-list-section-electric',
+    name: 'product-list-section-electric',
     dependencies: [],
-    registryDependencies: ['products-list-section'],
-    files: ['examples/sections/products-list-section/electric.tsx'],
-    component: lazy(() => import('./examples/sections/products-list-section/electric')),
+    registryDependencies: ['product-list-section'],
+    files: ['examples/sections/product-list-section/electric.tsx'],
+    component: lazy(() => import('./examples/sections/product-list-section/electric')),
   },
   {
-    name: 'products-list-section-warm',
+    name: 'product-list-section-warm',
     dependencies: [],
-    registryDependencies: ['products-list-section'],
-    files: ['examples/sections/products-list-section/warm.tsx'],
-    component: lazy(() => import('./examples/sections/products-list-section/warm')),
+    registryDependencies: ['product-list-section'],
+    files: ['examples/sections/product-list-section/warm.tsx'],
+    component: lazy(() => import('./examples/sections/product-list-section/warm')),
   },
   {
-    name: 'products-list-section-luxury',
+    name: 'product-list-section-luxury',
     dependencies: [],
-    registryDependencies: ['products-list-section'],
-    files: ['examples/sections/products-list-section/luxury.tsx'],
-    component: lazy(() => import('./examples/sections/products-list-section/luxury')),
+    registryDependencies: ['product-list-section'],
+    files: ['examples/sections/product-list-section/luxury.tsx'],
+    component: lazy(() => import('./examples/sections/product-list-section/luxury')),
   },
   {
     name: 'radio-group-example',
