+++ apps/web/vibes/soul/examples/sections/product-list-section/searchParams.ts
