+++ apps/web/vibes/soul/sections/product-list-section/filter-parsers.ts
@@ -1,6 +1,6 @@
 import { parseAsArrayOf, parseAsInteger, parseAsString, ParserBuilder } from 'nuqs/server';
 
-import { Filter } from './filters-panel';
+import { Filter } from './filter-panel';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function getFilterParsers(filters: Filter[]): Record<string, ParserBuilder<any>> {
