+++ apps/web/vibes/soul/sections/product-list-section/index.tsx
@@ -1,41 +1,42 @@
 import { Sliders } from 'lucide-react';
-import { Suspense } from 'react';
+import { ComponentProps } from 'react';
 
 import { Stream, Streamable } from '@/vibes/soul/lib/streamable';
 import { Button } from '@/vibes/soul/primitives/button';
 import { CursorPagination, CursorPaginationInfo } from '@/vibes/soul/primitives/cursor-pagination';
 import { ProductCardWithId } from '@/vibes/soul/primitives/product-card';
 import * as SidePanel from '@/vibes/soul/primitives/side-panel';
+import * as Skeleton from '@/vibes/soul/primitives/skeleton';
 import {
   Breadcrumbs,
   BreadcrumbsSkeleton,
   BreadcrumbWithId,
 } from '@/vibes/soul/sections/breadcrumbs';
 import { ProductList } from '@/vibes/soul/sections/product-list';
-import { Filter, FiltersPanel } from '@/vibes/soul/sections/products-list-section/filters-panel';
+import { Filter, FilterPanel } from '@/vibes/soul/sections/product-list-section/filter-panel';
 import {
   Sorting,
   SortingSkeleton,
   Option as SortOption,
-} from '@/vibes/soul/sections/products-list-section/sorting';
+} from '@/vibes/soul/sections/product-list-section/sorting';
 
-interface Props {
+export interface ProductListSectionProps {
   breadcrumbs?: Streamable<BreadcrumbWithId[]>;
-  title?: Streamable<string | null>;
+  title?: Streamable<string>;
   totalCount: Streamable<number>;
   products: Streamable<ProductCardWithId[]>;
   filters: Streamable<Filter[]>;
   sortOptions: Streamable<SortOption[]>;
   compareProducts?: Streamable<ProductCardWithId[]>;
   paginationInfo?: Streamable<CursorPaginationInfo>;
-  compareAction?: React.ComponentProps<'form'>['action'];
+  compareAction?: ComponentProps<'form'>['action'];
   compareLabel?: Streamable<string>;
   filterLabel?: string;
   filtersPanelTitle?: Streamable<string>;
   resetFiltersLabel?: Streamable<string>;
   rangeFilterApplyLabel?: Streamable<string>;
-  sortLabel?: Streamable<string | null>;
-  sortPlaceholder?: Streamable<string | null>;
+  sortLabel?: Streamable<string>;
+  sortPlaceholder?: Streamable<string>;
   sortParamName?: string;
   sortDefaultValue?: string;
   compareParamName?: string;
@@ -44,10 +45,27 @@ interface Props {
   placeholderCount?: number;
 }
 
-export function ProductsListSection({
+/**
+ * This component supports various CSS variables for theming. Here's a comprehensive list, along
+ * with their default values:
+ *
+ * ```css
+ * :root {
+ *   --product-list-section-title: hsl(var(--foreground));
+ *   --product-list-section-title-font-family: var(--font-heading);
+ *   --product-list-section-total-count: hsl(var(--contrast-300));
+ *   --product-list-section-filter-label-font-family: var(--font-mono);
+ *   --product-list-section-filter-label: hsl(var(--contrast-400));
+ *   --product-list-section-filter-link: hsl(var(--contrast-500));
+ *   --product-list-section-filter-link-hover: hsl(var(--foreground));
+ *   --product-list-section-filter-link-font-family: var(--font-body);
+ * }
+ * ```
+ */
+export function ProductListSection({
   breadcrumbs: streamableBreadcrumbs,
-  title = 'Products',
-  totalCount,
+  title: streamableTitle = 'Products',
+  totalCount: streamableTotalCount,
   products,
   compareProducts,
   sortOptions: streamableSortOptions,
@@ -67,9 +85,9 @@ export function ProductsListSection({
   emptyStateSubtitle,
   emptyStateTitle,
   placeholderCount = 8,
-}: Props) {
+}: ProductListSectionProps) {
   return (
-    <div className="group/products-list-section @container">
+    <div className="group/product-list-section @container">
       <div className="mx-auto max-w-screen-2xl px-4 py-10 @xl:px-6 @xl:py-14 @4xl:px-8 @4xl:py-12">
         <div>
           <Stream fallback={<BreadcrumbsSkeleton />} value={streamableBreadcrumbs}>
@@ -77,24 +95,23 @@ export function ProductsListSection({
               breadcrumbs && breadcrumbs.length > 1 && <Breadcrumbs breadcrumbs={breadcrumbs} />
             }
           </Stream>
-          <div className="flex flex-wrap items-center justify-between gap-4 pb-8 pt-6 text-foreground">
-            <h1 className="flex items-center gap-2 font-heading text-3xl font-medium leading-none @lg:text-4xl @2xl:text-5xl">
-              <Suspense
-                fallback={
-                  <span className="inline-flex h-[1lh] w-[6ch] animate-pulse rounded-lg bg-contrast-100" />
-                }
+          <div className="flex flex-wrap items-center justify-between gap-4 pb-8 pt-6">
+            <h1 className="flex items-center gap-2 font-[family-name:var(--product-list-section-title-font-family,var(--font-family-heading))] text-3xl font-medium leading-none text-[var(--product-list-section-title,hsl(var(--foreground)))] @lg:text-4xl @2xl:text-5xl">
+              <Stream fallback={<ProductListSectionTitleSkeleton />} value={streamableTitle}>
+                {(title) => <span>{title}</span>}
+              </Stream>
+              <Stream
+                fallback={<ProductListSectionTotalCountSkeleton />}
+                value={streamableTotalCount}
               >
-                {title}
-              </Suspense>
-              <Suspense
-                fallback={
-                  <span className="inline-flex h-[1lh] w-[2ch] animate-pulse rounded-lg bg-contrast-100" />
-                }
-              >
-                <span className="text-contrast-300">{totalCount}</span>
-              </Suspense>
+                {(totalCount) => (
+                  <span className="text-[var(--product-list-section-total-count,hsl(var(--contrast-300)))]">
+                    {totalCount}
+                  </span>
+                )}
+              </Stream>
             </h1>
-            <div className="flex gap-2">
+            <div className="flex items-center gap-2">
               <Stream
                 fallback={<SortingSkeleton />}
                 value={Streamable.all([
@@ -126,7 +143,7 @@ export function ProductsListSection({
                   <Stream value={streamableFiltersPanelTitle}>
                     {(filtersPanelTitle) => (
                       <SidePanel.Content title={filtersPanelTitle}>
-                        <FiltersPanel
+                        <FilterPanel
                           filters={filters}
                           paginationInfo={paginationInfo}
                           rangeFilterApplyLabel={rangeFilterApplyLabel}
@@ -145,7 +162,7 @@ export function ProductsListSection({
             <Stream value={streamableFiltersPanelTitle}>
               {(filtersPanelTitle) => <h2 className="sr-only">{filtersPanelTitle}</h2>}
             </Stream>
-            <FiltersPanel
+            <FilterPanel
               className="sticky top-4"
               filters={filters}
               paginationInfo={paginationInfo}
@@ -154,7 +171,7 @@ export function ProductsListSection({
             />
           </aside>
 
-          <div className="flex-1 group-has-[[data-pending]]/products-list-section:animate-pulse">
+          <div className="flex-1 group-has-[[data-pending]]/product-list-section:animate-pulse">
             <ProductList
               compareAction={compareAction}
               compareLabel={compareLabel}
@@ -172,5 +189,27 @@ export function ProductsListSection({
         </div>
       </div>
     </div>
+  );
+}
+
+export function ProductListSectionTitleSkeleton() {
+  return (
+    <Skeleton.Root
+      className="group-has-[[data-pending]]/product-list-section:animate-pulse"
+      pending
+    >
+      <Skeleton.Text characterCount={6} className="rounded-lg" data-pending />
+    </Skeleton.Root>
+  );
+}
+
+export function ProductListSectionTotalCountSkeleton() {
+  return (
+    <Skeleton.Root
+      className="group-has-[[data-pending]]/product-list-section:animate-pulse"
+      pending
+    >
+      <Skeleton.Text characterCount={2} className="rounded-lg" data-pending />
+    </Skeleton.Root>
   );
 }
